+++ ui/app/templates/vault/cluster/secrets/backend/transit-actions-layout.hbs
@@ -51,19 +51,12 @@
         </ul>
       </nav>
     </div>
-    <TransitKeyActionsTwo
+    <TransitKeyActions
       @selectedAction={{this.selectedAction}}
       @backend={{this.backend}}
       @key={{this.model}}
       @capabilities={{this.capabilities}}
       @onRefresh={{action "refresh"}}
     />
-    {{!-- <TransitKeyActions
-      @selectedAction={{this.selectedAction}}
-      @backend={{this.backend}}
-      @key={{this.model}}
-      @capabilities={{this.capabilities}}
-      @onRefresh={{action "refresh"}}
-    /> --}}
   </div>
 </div>
